Stricter paths to prevent traversal & remove custom tempDir option

+++ index.js
@@ -12,34 +12,35 @@ export default async function downloader(downloadUrl, checksumUrl, options) {
   // intialize options if none are set
   options = options || {};
 
-  // don't delete the temp dir if set manually and dir exists
-  let deleteTempDir = true;
-  if (options.tempDir && fs.pathExistsSync(options.tempDir)) {
-    deleteTempDir = false;
-  }
-
   // normalize options and set defaults
   options = {
     filename: options.filename || urlParse(downloadUrl).pathname.split("/").pop(),
     extract: !!options.extract,
-    tempDir: options.tempDir ? path.resolve(process.cwd(), options.tempDir) : tempy.directory(),
-    destDir: options.destDir ? path.resolve(process.cwd(), options.destDir) : path.resolve(process.cwd(), "download"),
+    destDir: options.destDir ? path.resolve(process.cwd(), options.destDir) : path.resolve(process.cwd(), "downloads"),
     cleanDestDir: !!options.cleanDestDir,
     algorithm: options.algorithm || "sha256",
     encoding: options.encoding || "binary",
   };
 
+  // throw an error if destDir is outside of the module to prevent path traversal for security reasons
+  if (!options.destDir.startsWith(process.cwd())) {
+    throw new Error(`destDir must be located within '${process.cwd()}', it's currently set to '${options.destDir}'.`);
+  }
+
+  // initialize temporary directory
+  const tempDir = tempy.directory();
+
   try {
     // simultaneously download the desired file and its checksums
     await Promise.all([
-      downloadFile(downloadUrl, path.join(options.tempDir, options.filename)),
-      downloadFile(checksumUrl, path.join(options.tempDir, "checksums.txt")),
+      downloadFile(downloadUrl, path.join(tempDir, options.filename)),
+      downloadFile(checksumUrl, path.join(tempDir, "checksums.txt")),
     ]);
 
     // validate the checksum of the download
-    if (await checkChecksum(options.tempDir, options.filename, "checksums.txt", options.algorithm, options.encoding)) {
+    if (await checkChecksum(tempDir, options.filename, "checksums.txt", options.algorithm, options.encoding)) {
       // optionally clear the target directory of existing files
-      if (options.cleanDestDir) {
+      if (options.cleanDestDir && fs.existsSync(options.destDir)) {
         await fs.remove(options.destDir);
       }
 
@@ -48,21 +49,19 @@ export default async function downloader(downloadUrl, checksumUrl, options) {
 
       if (options.extract) {
         // decompress download and move resulting files to final destination
-        await decompress(path.join(options.tempDir, options.filename), options.destDir);
+        await decompress(path.join(tempDir, options.filename), options.destDir);
         return options.destDir;
       } else {
         // move verified download to final destination as-is
-        await fs.copy(path.join(options.tempDir, options.filename), path.join(options.destDir, options.filename));
+        await fs.copy(path.join(tempDir, options.filename), path.join(options.destDir, options.filename));
         return path.join(options.destDir, options.filename);
       }
     } else {
       throw new Error(`Invalid checksum for ${options.filename}.`);
     }
   } finally {
-    // delete temporary directory (except for edge cases above)
-    if (deleteTempDir) {
-      await fs.remove(options.tempDir);
-    }
+    // delete temporary directory
+    await fs.remove(tempDir);
   }
 }
